+++ modules/apps/rollerRework/constants.js
@@ -1,0 +1,6 @@
+export const actionLabels = [ 
+    {side:"action", quadrant: 1, label: "Action Setup"}
+    , {side:"action", quadrant: 2, label: "Action"}
+    , {side:"reaction", quadrant: 3, label: "Reaction Setup"}
+    , {side:"reaction", quadrant: 4, label: "Reaction"}
+]