+++ modules/sheets/stand-actor-sheet.js
@@ -6,7 +6,7 @@ export class StandSheet extends BaseActorSheet {
 	static get defaultOptions() {
 		return foundry.utils.mergeObject(super.defaultOptions, {
 			classes: ["bizarre-adventures-d6", "sheet", "actor", "stand"]
-			, template: "systems/bizarre-adventures-d6/templates/sheets/stand-actor-sheet.hbs"
+			, template: "systems/bizarre-adventures-d6/templates/actor/sheets/stand-actor-sheet.hbs"
 			, width: 800
 			, height: 800
 			, tabs: [{